refactor(oas): extract `VendorExamples`
closes #224
