+++ packages/openapi-sampler/src/traverse/index.ts
@@ -1,2 +1,3 @@
 export * from './Traverse';
 export * from './DefaultTraverse';
+export * from './VendorExamples';
