+++ packages/openapi-sampler/tests/example.spec.ts
@@ -99,253 +99,4 @@ describe('Example', () => {
     // assert
     expect(result).toEqual('foo');
   });
-
-  it.each([
-    {
-      [VendorExtensions.X_EXAMPLE]: {
-        name: 'name',
-        age: 30
-      }
-    },
-    {
-      [VendorExtensions.X_EXAMPLE]: {
-        'application/json': {
-          name: 'name',
-          age: 30
-        }
-      }
-    },
-    {
-      [VendorExtensions.X_EXAMPLES]: {
-        'application/json': {
-          name: 'name',
-          age: 30
-        }
-      }
-    },
-    {
-      [VendorExtensions.X_EXAMPLES]: {
-        'application/json': {
-          some: {
-            name: 'name',
-            age: 30
-          }
-        }
-      }
-    },
-    {
-      [VendorExtensions.X_EXAMPLES]: {
-        'application/json': {
-          some: {
-            name: 'name',
-            age: 30
-          }
-        }
-      }
-    },
-    {
-      [VendorExtensions.X_EXAMPLES]: [
-        {
-          name: 'name',
-          age: 30
-        }
-      ]
-    },
-    {
-      [VendorExtensions.X_EXAMPLES]: {
-        'application/json': [
-          {
-            name: 'name',
-            age: 30
-          }
-        ]
-      }
-    }
-  ])(
-    'should match %j object vendor example when includeVendorExamples is true',
-    (input) => {
-      // arrange
-      const expected = {
-        name: 'name',
-        age: 30
-      };
-
-      const schema = {
-        type: 'object',
-        properties: {
-          name: {
-            type: 'string'
-          },
-          age: {
-            type: 'integer'
-          }
-        },
-        ...input
-      };
-
-      // act
-      const result = sample(schema, { includeVendorExamples: true });
-
-      // assert
-      expect(result).toMatchObject(expected);
-    }
-  );
-
-  it.each(['some-string', { name: 'some-name', points: 100 }])(
-    'should ignore %j object vendor example when includeVendorExamples is true',
-    (input) => {
-      // arrange
-      const schema = {
-        type: 'object',
-        properties: {
-          name: {
-            type: 'string'
-          },
-          age: {
-            type: 'integer'
-          }
-        },
-        [VendorExtensions.X_EXAMPLES]: {
-          'some-example': input
-        }
-      };
-
-      // act
-      const result = sample(schema, { includeVendorExamples: true });
-
-      // assert
-      expect(result).toMatchObject({ age: 42, name: 'lorem' });
-    }
-  );
-
-  it.each([
-    [
-      {
-        [VendorExtensions.X_EXAMPLE]: [
-          {
-            name: 'name',
-            age: 30
-          }
-        ]
-      },
-      {
-        [VendorExtensions.X_EXAMPLE]: {
-          'application/json': [
-            {
-              name: 'name',
-              age: 30
-            }
-          ]
-        }
-      },
-      {
-        [VendorExtensions.X_EXAMPLES]: {
-          'application/json': [
-            {
-              name: 'name',
-              age: 30
-            }
-          ]
-        }
-      },
-      {
-        [VendorExtensions.X_EXAMPLES]: {
-          'application/json': [
-            {
-              some: {
-                name: 'name',
-                age: 30
-              }
-            }
-          ]
-        }
-      },
-      {
-        [VendorExtensions.X_EXAMPLES]: [
-          {
-            name: 'name',
-            age: 30
-          }
-        ]
-      },
-      {
-        [VendorExtensions.X_EXAMPLES]: {
-          'application/json': [
-            {
-              name: 'name',
-              age: 30
-            }
-          ]
-        }
-      }
-    ]
-  ])(
-    'should match %j array vendor example when includeVendorExamples is true',
-    (input) => {
-      // arrange
-      const expected = [
-        {
-          name: 'name',
-          age: 30
-        }
-      ];
-
-      const schema = {
-        type: 'array',
-        items: {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'string'
-            },
-            age: {
-              type: 'integer'
-            }
-          }
-        },
-        ...input
-      };
-
-      // act
-      const result = sample(schema, { includeVendorExamples: true });
-
-      // assert
-      expect(result).toMatchObject(expected);
-    }
-  );
-
-  it.each([
-    [[{ name: 'some-name', age: 30 }]],
-    2,
-    'some-string',
-    { name: 'some-name', points: 100 }
-  ])(
-    'should ignore %j array vendor example when includeVendorExamples is true',
-    (input) => {
-      // arrange
-      const schema = {
-        type: 'array',
-        items: {
-          type: 'object',
-          properties: {
-            name: {
-              type: 'string'
-            },
-            age: {
-              type: 'integer'
-            }
-          }
-        },
-        [VendorExtensions.X_EXAMPLES]: {
-          'some-example': input
-        }
-      };
-
-      // act
-      const result = sample(schema, { includeVendorExamples: true });
-
-      // assert
-      expect(result).toMatchObject([{ age: 42, name: 'lorem' }]);
-    }
-  );
 });
